(feat) Add Xirr for individual fund returns

+++ src/main/java/com/invest/indices/domain/model/ReturnInputs.java
@@ -5,7 +5,7 @@ import lombok.Data;
 
 @Data
 public class ReturnInputs {
-    private int invAmount;
+    private Double invAmount;
     private int schemeCode;
     private String fromDate;
     private String toDate;
